Accept manual PATH and trim the input before use
When setting PATH directory manually, because the automatic detection
failed, bin enters a loop and never comes out of it. This change adds a
break after accepting the input and breaks out of the loop.

Also, the input seems to contain a newline at the end, which causes
further installation of binaries to fail due to malformed binary path.
In order to fix it, trimp white spaces from the given input before using
it.