make ensure check file hash
fixes #205

this is a breaking change because since everyone running `bin ensure`
after this gets merged will result in all their binaries getting
re-downloaded. This is good though since that will fix all the incorrect
hashes that are wrong in the first place.

I've decided not to add a flag to skip this check since checking hashes
is the right thing to do in the first place.

Signed-off-by: Marcos Lilljedahl <marcosnils@gmail.com>